Small organizational cleanup
