+++ services/client/src/api/types.ts
@@ -1,13 +1,10 @@
-import { OrderDirection, PoolOrderByEnum } from '../utils';
-
 export interface FetchHypesFilter {
   first: number;
   skip: number;
   text?: string;
-  // orderBy?: PoolOrderByEnum;
-  // order?: OrderDirection;
 }
 
 export const API = `${process.env.REACT_APP_API_HOST}`;
 export const AUTH_API = `${process.env.REACT_APP_WALLET_HOST}`;
 export const GRAPHQL_API = `${process.env.REACT_APP_GRAPHQL_HOST}`;
+export const TelegramBotName = process.env.REACT_APP_TELEGRAM_BOT_NAME || 'HypeAppBot';
