--- services/client/src/config.ts
@@ -1,1 +1,0 @@
-export const TelegramBotName = process.env.REACT_APP_TELEGRAM_BOT_NAME || 'HypeAppBot';
