+++ services/client/src/models/index.ts
@@ -2,3 +2,5 @@ export * from './HypePool.model';
 export * from './LoginSignature.model';
 export * from './User.model';
 export * from './JWTResponse.model';
+export * from './Redeem.model';
+export * from './Reward.model';
