+++ services/client/src/pages/Home/Home.tsx
@@ -3,7 +3,7 @@ import Megaphone from '../../assets/images/megaphone.png';
 import Input from '../../components/input/Input';
 import Image from '../../components/image/Image';
 import LoadingSpinner from '../../assets/icons/Spinner';
-import NotFoundIcon from 'src/assets/icons/NotFound';
+import NotFoundIcon from '../../assets/icons/NotFound';
 import { useHomeEffects } from './Home.effects';
 import {
   HeroContainer,
